test: add suffix to filename for compare with snapshot
